backend/requirements.txt
flask==2.3.2
flask-socketio==5.3.6
paho-mqtt==1.6.1
python-dotenv==1.0.0
gunicorn==20.1.0
eventlet==0.33.3